+++ packages/sync-rules/src/compiler/sqlite.ts
@@ -19,15 +19,38 @@ import {
 } from '../sync_plan/expression.js';
 import { ConnectionParameterSource } from '../sync_plan/plan.js';
 import { ParsingErrorListener } from './compiler.js';
-import { SourceResultSet } from './table.js';
+import { BaseSourceResultSet, PhysicalSourceResultSet, SourceResultSet, SyntacticResultSetSource } from './table.js';
+import { SqlScope } from './scope.js';
+import { SourceSchema } from '../types.js';
 
 export interface ResolvedSubqueryExpression {
   filters: SqlExpression<ExpressionInput>[];
   output: SqlExpression<ExpressionInput>;
 }
 
+/**
+ * A prepared subquery or common table expression.
+ */
+export interface PreparedSubquery {
+  /**
+   * Columns selected by the query, indexed by their name.
+   */
+  resultColumns: Record<string, SqlExpression<ExpressionInput>>;
+
+  /**
+   * Tables the subquery selects from.
+   */
+  tables: Map<SyntacticResultSetSource, SourceResultSet>;
+
+  /**
+   * Filters affecting the subquery.
+   */
+  where: SqlExpression<ExpressionInput> | null;
+}
+
 export interface PostgresToSqliteOptions {
   readonly originalText: string;
+  readonly scope: SqlScope;
   readonly errors: ParsingErrorListener;
   readonly locations: NodeLocations;
 
@@ -36,7 +59,7 @@ export interface PostgresToSqliteOptions {
    *
    * Should report an error if resolving the table failed, using `node` as the source location for the error.
    */
-  resolveTableName(node: ExprRef, name: string | nil): SourceResultSet | null;
+  resolveTableName(node: ExprRef, name: string | nil): SourceResultSet | PreparedSubquery | null;
 
   /**
    * Generates a table alias for synthetic subqueries like those generated to desugar `IN` expressions to `json_each`
@@ -104,11 +127,38 @@ export class PostgresToSqlite {
           return this.invalidExpression(expr, '* columns are not supported here');
         }
 
-        const instantiation = new ColumnInRow(expr, resultSet, expr.name);
-        return {
-          type: 'data',
-          source: instantiation
-        };
+        // If this references something from a source table, warn if that column doesn't exist.
+        if (resultSet instanceof PhysicalSourceResultSet && resultSet.schemaTablesForWarnings.length) {
+          let columnExistsInAnySourceTable = false;
+
+          for (const table of resultSet.schemaTablesForWarnings) {
+            if (table.getColumn(expr.name) != null) {
+              columnExistsInAnySourceTable = true;
+              break;
+            }
+          }
+
+          if (!columnExistsInAnySourceTable) {
+            this.options.errors.report('Column not found.', expr, { isWarning: true });
+          }
+        }
+
+        if (resultSet instanceof BaseSourceResultSet) {
+          // This is an actual result set.
+          const instantiation = new ColumnInRow(expr, resultSet, expr.name);
+          return {
+            type: 'data',
+            source: instantiation
+          };
+        } else {
+          // Resolved to a subquery, inline the reference.
+          const expression = resultSet.resultColumns[expr.name];
+          if (expression == null) {
+            return this.invalidExpression(expr, 'Column not found in subquery.');
+          }
+
+          return expression;
+        }
       }
       case 'parameter':
         return this.invalidExpression(
@@ -145,11 +195,12 @@ export class PostgresToSqlite {
           return this.invalidExpression(expr.function, 'DISTINCT, ORDER BY, FILTER and OVER clauses are not supported');
         }
 
-        const forbiddenReason = forbiddenFunctions[expr.function.name];
+        const functionName = expr.function.name.toLowerCase();
+        const forbiddenReason = forbiddenFunctions[functionName];
         if (forbiddenReason) {
           return this.invalidExpression(expr.function, `Forbidden call: ${forbiddenReason}`);
         }
-        let allowedArgs = supportedFunctions[expr.function.name];
+        let allowedArgs = supportedFunctions[functionName];
         if (allowedArgs == null) {
           return this.invalidExpression(expr.function, 'Unknown function');
         } else {
@@ -171,7 +222,7 @@ export class PostgresToSqlite {
 
         return {
           type: 'function',
-          function: expr.function.name,
+          function: functionName,
           parameters: expr.args.map((a) => this.translateNodeWithLocation(a))
         };
       }
@@ -183,6 +234,9 @@ export class PostgresToSqlite {
         const left = this.translateNodeWithLocation(expr.left);
         const right = this.translateNodeWithLocation(expr.right);
         if (expr.op === 'LIKE') {
+          // We don't support LIKE in the old bucket definition system, and want to make sure we're clear about ICU,
+          // case sensitivity and changing the escape character first. TODO: Support later.
+          this.options.errors.report('LIKE expressions are not currently supported.', expr);
           return { type: 'function', function: 'like', parameters: [left, right] };
         } else if (expr.op === 'NOT LIKE') {
           return {
@@ -210,8 +264,9 @@ export class PostgresToSqlite {
         let rightHandSideOfIs: SqlExpression<ExpressionInput>;
 
         switch (expr.op) {
-          case '+':
           case '-':
+            return this.invalidExpression(expr, 'Unary minus is not currently supported');
+          case '+':
             return { type: 'unary', operator: expr.op, operand: this.translateNodeWithLocation(expr.operand) };
           case 'NOT':
             return { type: 'unary', operator: 'not', operand: this.translateNodeWithLocation(expr.operand) };
@@ -314,6 +369,14 @@ export class PostgresToSqlite {
     } else if (right.type == 'call' && right.function.name.toLowerCase() == 'row') {
       return this.desugarInValues(negated, expr.left, right.args);
     } else {
+      if (right.type == 'ref' && right.table == null) {
+        const cte = this.options.scope.resolveCommonTableExpression(right.name);
+        if (cte) {
+          // Something of the form x IN $cte.
+          return this.desugarInCte(negated, expr, right.name, cte);
+        }
+      }
+
       return this.desugarInScalar(negated, expr, right);
     }
   }
@@ -369,6 +432,28 @@ export class PostgresToSqlite {
       type: 'select',
       columns: [{ expr: { type: 'ref', name: 'value', table: { name } } }],
       from: [{ type: 'call', function: { name: 'json_each' }, args: [right], alias: { name } }]
+    });
+  }
+
+  /**
+   * Desugar `$left IN cteName` to `$left IN (SELECT cteName.onlyColumn FROM cteName)`.
+   */
+  private desugarInCte(
+    negated: boolean,
+    binary: ExprBinary,
+    cteName: string,
+    cte: PreparedSubquery
+  ): SqlExpression<ExpressionInput> {
+    const columns = Object.keys(cte.resultColumns);
+    if (columns.length != 1) {
+      return this.invalidExpression(binary.right, 'Common-table expression must return a single column');
+    }
+
+    const name = columns[0];
+    return this.desugarInSubquery(negated, binary, {
+      type: 'select',
+      columns: [{ expr: { type: 'ref', name, table: { name: cteName } } }],
+      from: [{ type: 'table', name: { name: cteName } }]
     });
   }